requirements.txt
python-chess==1.11.2
torch==2.11.0
numpy==2.2.6
scikit-learn==1.7.1
flask==3.0.0
